introduce sync monitoring concept
