+++ novawallet/Common/Services/CloudBackup/CloudBackupServiceFactoryProtocol.swift
@@ -3,7 +3,6 @@ import SoraKeystore
 
 protocol CloudBackupServiceFactoryProtocol {
     func createAvailabilityService() -> CloudBackupAvailabilityServiceProtocol
-    func createStorageManager(for baseUrl: URL) -> CloudBackupStorageManaging
     func createOperationFactory() -> CloudBackupOperationFactoryProtocol
     func createFileManager() -> CloudBackupFileManaging
     func createCodingManager() -> CloudBackupCoding
@@ -11,5 +10,4 @@ protocol CloudBackupServiceFactoryProtocol {
     func createDiffCalculator() -> CloudBackupDiffCalculating
     func createSecretsExporter(from keychain: KeystoreProtocol) -> CloudBackupSecretsExporting
     func createSecretsImporter(to keychain: KeystoreProtocol) -> CloudBackupSecretsImporting
-    func createUploadFactory() -> CloudBackupUploadFactoryProtocol
 }
