+++ novawallet/Common/Services/CloudBackup/Model/CloudBackupFileManaging.swift
@@ -3,7 +3,6 @@ import Foundation
 protocol CloudBackupFileManaging {
     func getFileName() -> String
     func getFileUrl() -> URL?
-    func getTempUrl() -> URL?
     func getBaseUrl() -> URL?
 }
 
@@ -29,10 +28,5 @@ extension ICloudBackupFileManager: CloudBackupFileManaging {
     func getFileUrl() -> URL? {
         let baseUrl = getBaseUrl() as? NSURL
         return baseUrl?.appendingPathComponent(getFileName())
-    }
-
-    func getTempUrl() -> URL? {
-        let baseUrl = getBaseUrl() as? NSURL
-        return baseUrl?.appendingPathComponent(CloudBackup.walletsTempFilename)
     }
 }
