--- novawallet/Common/Services/CloudBackup/Monitor/CloudBackupUpdateMonitoring.swift
@@ -1,20 +1,0 @@
-import Foundation
-
-enum CloudBackupUpdateMonitorError: Error {
-    case internalError(Error)
-}
-
-enum CloudBackupUpdateStatus {
-    case noFile
-    case downloaded
-    case notDownloaded
-    case unknown
-}
-
-typealias CloudBackupUpdateMonitoringClosure = (Result<CloudBackupUpdateStatus, CloudBackupUpdateMonitorError>) -> Void
-
-protocol CloudBackupUpdateMonitoring {
-    func start(notifyingIn queue: DispatchQueue, with closure: @escaping CloudBackupUpdateMonitoringClosure)
-
-    func stop()
-}
